Updated DataSet to combine Tensors instead of using TensorArray.
