+++ Sources/MimicTransferables/Types/DeviceType.swift
