+++ Sources/MimicTransferables/Types/Graph.swift
